Retreive streaming data for 256 kbps AAC Adding the parameters c=WEB_REMIX and cver=0.1 to the get_video_info request in get_song() returns a streamingData adaptiveFormat of itag 141, audioQuality of AUDIO_QUALITY_HIGH, a 256 kbps AAC m4a for authenticated requests from YouTube Music Premium subscribers.
Unfortunately, these parameters result in
player_response['videoDetails']['shortDescription'] not being returned,
the information in which contains useful song metadata.

player_response['microformat']['microformatDataRenderer']['description']
is returned, but it seems to be a truncated version of shortDescription,
so the metadata parsing in get_song() does not work on that response.

Therefore, a second get_video_info request is made with the updated
parameters and the metadata in song_meta is copied in to the new
player_response['videoDetails'], which is returned from get_song()
instead of song_meta.