get_artist: Don't fail on channels with only one category of content
The current check is meant to detect channels with only video content,
but it also hits small bands that for example have only released singles
so far. An example for this is "UC1uK8RT3m4nNpXmOSsFsZeg".

Since failing is not really helpful anyway (for example in audiotube, it
could still play only the sound of a channel that contains only videos),
I propose to remove the check altogether.