nav() now displays some debugging information on KeyError
There's a user reporting a KeyError in nav(), so I've changed it to
print out the key and dict for future debugging.  Also, a bare
Exception was used to capture a dictionary failure, removed that
since my change checks for the existance of the key in the dictionary.
If this is objectionable at all, you probably should change it to
catching a KeyError rather than catching Exception.